Moving health check to a different class
